+++ src/Exceptions/StaticallException.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace CodeDistortion\Staticall\Exceptions;
+
+use Exception;
+
+/**
+ * The main Staticall exception class.
+ */
+class StaticallException extends Exception
+{
+}